Make tx index commands
