+++ osmoutils/osmocli/index_cmd.go
@@ -1,0 +1,31 @@
+package osmocli
+
+import (
+	"fmt"
+
+	"github.com/spf13/cobra"
+)
+
+// Index command, but short is not set. That is left to caller.
+func IndexCmd(moduleName string) *cobra.Command {
+	return &cobra.Command{
+		Use:                        moduleName,
+		Short:                      fmt.Sprintf("Querying commands for the %s module", moduleName),
+		DisableFlagParsing:         true,
+		SuggestionsMinimumDistance: 2,
+		RunE:                       indexRunCmd,
+	}
+}
+
+func indexRunCmd(cmd *cobra.Command, args []string) error {
+	usageTemplate := `Usage:{{if .HasAvailableSubCommands}}
+  {{.CommandPath}} [command]{{end}}
+  
+{{if .HasAvailableSubCommands}}Available Commands:{{range .Commands}}{{if .IsAvailableCommand}}
+  {{rpad .Name .NamePadding }} {{.Short}}{{end}}{{end}}
+
+Use "{{.CommandPath}} [command] --help" for more information about a command.{{end}}
+`
+	cmd.SetUsageTemplate(usageTemplate)
+	return cmd.Help()
+}
